# Issue 1842: Major Performance bottleneck in Workflow Step at significant XHRLoad cost.